Member 3: Pao-Yu Chen
User Story 1: As a player, I want to hear victory/game over music , so that the players feel a sense of game completion and a heightened sense of excitement for the winning player. (30mins) 

User Story 2: As a player, I want to be able to see a victory, or defeat visual when i win or lose, so that to get visual clarification on the overall result of this fight. (1h) 

User Story 3: As a player, I want to be able to easily restart the game (e.g. via UI button or Keyboard Key to restart) after the game-over sequence, so that I can easily play again without having to close and reopen the application. (1h30mins) 

User Story 4: As a player, I want to see my battle stats (i.e. number of successful hits & blocks) after the battle ends, so that I can gain an overall understanding of my game performance. (1h)